Movie list fix

+++ app/src/main/java/com/example/bookingmovie/data/Movie/MovieEntity.kt
@@ -6,12 +6,12 @@ import androidx.room.PrimaryKey
 data class MovieEntity(
     @PrimaryKey(autoGenerate = true) val movie_id: Int = 0,
     val movie_name: String,
-    val genre_id : Int,
-    val showtime_id : Int,
-    val description :String,
-    val price : Float,
-    val banner : String?,
-    val video : String?,
+    val genre_id: Int,
+    val showtime_id: Int,
+    val description:String,
+    val price: Double,
+    val banner: String?,
+    val video: String?,
     val year: Int
 )
